Pull up following revision(s) (requested by manu in ticket #607): 	lib/libc/arch/i386/gen/isnanl.c: revision 1.6 	lib/libc/gdtoa/gdtoa.c: revision 1.2-1.3 	lib/libc/arch/x86_64/gen/isnanl.c: revision 1.6 	lib/libc/gdtoa/gdtoaimp.h: revision 1.6 	sys/arch/m68k/include/ieee.h: revision 1.13 	usr.bin/xlint/lint1/scan.l: revision 1.36-1.37 	lib/libc/stdio/snprintf_ss.c: revision 1.4 	lib/libc/arch/i386/gen/isfinitel.c: revision 1.2 	lib/libc/stdio/vfscanf.c: revision 1.38 	sys/arch/sparc/include/ieee.h: revision 1.11-1.12 	lib/libc/gdtoa/dtoa.c: revision 1.4 	lib/libc/stdio/Makefile.inc: revision 1.35 	lib/libc/stdio/fvwrite.c: revision 1.17 	lib/libc/arch/m68k/gen/fpclassifyl.c: revision 1.2 	lib/libc/arch/i386/gen/isinfl.c: revision 1.6 	lib/libc/arch/x86_64/gen/isinfl.c: revision 1.6 	lib/libc/arch/x86_64/gen/isfinitel.c: revision 1.2 	lib/libc/stdio/vfprintf.c: revision 1.55-1.57 	lib/libc/stdio/vsnprintf_ss.c: revision 1.3 	lib/libc/stdio/vfwprintf.c: revision 1.10 	sys/arch/x86/include/ieee.h: revision 1.10 	lib/libc/gdtoa/dmisc.c: revision 1.3 	lib/libc/gdtoa/Makefile.inc: revision 1.5 	sys/arch/hppa/include/ieee.h: revision 1.10 	lib/libc/arch/x86_64/gen/fpclassifyl.c: revision 1.3 	lib/libc/arch/i386/gen/fpclassifyl.c: revision 1.2 	sys/sys/ieee754.h: revision 1.7 	lib/libc/gdtoa/gdtoa.h: revision 1.7 	include/stdio.h: revision 1.67-1.68 	lib/libc/gdtoa/hdtoa.c: revision 1.1-1.4 	lib/libc/gdtoa/ldtoa.c: revision 1.1-1.4 defined(_NETBSD_SOURCE) is equivalent to (!defined(_ANSI_SOURCE) && !defined(_POSIX_C_SOURCE) && !defined(_XOPEN_SOURCE)), so there's no need to check both of them. Fix for issue reported in PR lib/35401 as well as related overflow bugs. deal with hex doubles. Instead of abusing stdio to get a signal-safe version of sprintf, provide one. remove __SAFE add long double and hex double support from freebsd. make this compile. add new prototypes. add the new files to the build. Note I am not bumping libc now, because these are not used yet. Merge the int bit with the high fraction bit. Add constants/macros needed by gdtoa. add constants used by gdtoa since the int bit is merged, do the explicit math. ext_int bit is no more. ext_int bit is no more. - merge change from freebsd - add support for building as vfprintf.c - XXX: we strdup to simplify the freeing logic. This should be fixed for   efficiency in the vfprintf case. use vfwprintf.c enable wide doubles. some int -> size_t deal with sparc64 that has 112 bits of mantissa. make extended precision gdtoa friendly. int/size_t changes make this gdtoa friendly. remove dup definition use dtoa() instead of returning empty when we don't have extended precision information. Fix previous, add forgotten pointer dereference in the call to dtoa(). Add a cheesy workaround marked XXX for the situation where the strtod() implementation available in the environment does not handle hex floats. Discussed with and suggested by christos From Christos:  gdtoa fixes for m68k.  M68k ports should build now, but printing extended precision is a little off. vax does not have <machine/ieee.h> or long double It would be nice if the compiler provided something like __IEEE_MATH__ bring in FreeBSD's vfscanf() to gain multi-byte/collation support. Unfortunately it is too difficult to make vfwscanf and this share the same code like I did with printf, because for string parsing the code is too different.

+++ lib/libc/gdtoa/dtoa.c
@@ -1,4 +1,4 @@
-/* $NetBSD: dtoa.c,v 1.3 2006/02/28 18:51:08 kleink Exp $ */
+/* $NetBSD: dtoa.c,v 1.3.4.1 2007/05/07 19:49:06 pavel Exp $ */
 
 /****************************************************************
 
@@ -317,7 +317,7 @@ dtoa
 			if (i <= 0)
 				i = 1;
 		}
-	s = s0 = rv_alloc(i);
+	s = s0 = rv_alloc((size_t)i);
 
 #ifdef Honor_FLT_ROUNDS
 	if (mode > 1 && rounding != 1)
